Fix some potentially unstable sigs in PlayerState
This commit purely consists of some fixups of some potentially-problematic signatures that may break. Thanks to Aers and Pohky for catching these and informing me about them.

The IsMcGuffinUnlocked signature relied on a few offsets remaining stable, which probably won't happen, but no point in tempting fate.

The IsFramersKitUnlocked signature had a hardcoded 0x28 in it, which would have broken upon SE adding a new Framer's Kit, so the signature has been extended a bit.