BGCollision Math classes
BGCollision uses its own set of math classes with their own set of functions (in the client).

Normally it would be fine to use the common math classes but the BGCollision Vector3 is smaller than the regular Vector3.

Some math classes were duplicated. Not happy with this but its a quick fix so things don't break for people.