BoneSimulator: Add Spring field
Discovered this when messing around with the bone simulator and think it's worth adding.

This is a weird field, but messing around with it and reading the disassembly it seems to be a "springiness" value. It starts a little lower than 60.0, but anything higher than that results in relentless jitter. Lower values such as 0.0 completely remove the springiness/bounce.