Add AtkClippingMaskNode This is essentially an image node with a monochrome texture loaded, which serves as a mask for other nodes in the draw list.
Every example I've found of this type has been located within an `AtkComponentNode`, rather than being in the Addon's main node list. The mask is seemingly applied to the `NextSiblingNode` and to all subsequent siblings within the Component. It is not applied to previous sibling nodes, or to nodes outside the Component.