AgentCharaCard: Fix reading decorations
I'm a total idiot and only tested this code against a full adventurer plate, and it turns out almost everything I did was completely wrong. I assumed the location of each decoration is static, but it turns out this structure is more extendable than I thought, I guess to support more decoration types in the future.

The decoration data is stored in two different fixed size arrays with five elements (since there's only five decoration types.) The first array is the row indices, where the hard-coded locations were before. The second array is some more decoration information, but currently I only know of and care about the type for now.

This does not affect base plates or borders, which are not considered decorations.