Add CanFly field to PlayerState
This is a boolean set while processing the InitZone packet, and
informs if the player is authorized to fly in said zone. I replaced
the old comment referring to something else which doesn't seem to
occupy that position anymore, and also added where to find the
function that sets this field.